Modularize WASM bindings into namespaced submodules
Split wasm.rs (4673 lines) into separate files by functionality:
- cell, composition, defects, elastic, element, io, lattice
- md, neighbors, optimizers, order_params, potentials
- properties, species, structure, surfaces, symmetry
- trajectory, transformations, xrd

Mirrors the Python bindings modularization for consistency.
